chore: rename api/sentence-collector to api/pending-sentences
